% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakUpCluster.R
\name{BreakUpCluster}
\alias{BreakUpCluster}
\title{BreakUpCluster}
\usage{
BreakUpCluster(Object = Object, breakup.cluster = 6, sub.cluster = 3)
}
\arguments{
\item{Object}{A PathwayObject}

\item{breakup.cluster}{An integer with the number of the cluster to be broken up}

\item{sub.cluster}{The amount of subclusters to be generated}
}
\value{
PathwayObject
}
\description{
BreakUpCluster
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
man.Great.Object1 <- ManageGeneSets(Object = Great.bckgnrd.Object1,
                                   keep.type =c("Disease Ontology",
                                   "GO Biological Process" ),
                                   exclude.type="")
man.Great.Object2 <- CombineGeneSets(Object = man.Great.Object1)
man.Great.Object3 <- ClusterGeneSets(Object = man.Great.Object2,
                                     clusters = 5,
                                     method = "kmeans")
                                     
man.Great.Object3 <- (Object = man.Great.Object3, breakup.cluster = 3, sub.cluster=3)                                     
                                     
}
