% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterGeneSets.R
\name{ClusterGeneSets}
\alias{ClusterGeneSets}
\title{ClusterGeneSets}
\usage{
ClusterGeneSets(Object, clusters = 5, method = "kmeans",
  order = "group", molecular.signature = "All")
}
\arguments{
\item{Object}{A PathwayObject}

\item{clusters}{A numeric with the number of clusters required}

\item{method}{The cluster method specified, Options are kmeans, kmeans_group,
Hierarchical and Hierarchical_group}

\item{order}{How should the data be ordered, by group or by cluster}

\item{molecular.signature}{to filter the molecular signatures which are duplicated
before clustering. options are Unique or All}
}
\value{
PathwayObject
}
\description{
ClusterGeneSets
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
man.Great.Object1 <- ManageGeneSets(Object = Great.bckgnrd.Object1,
                                   keep.type =c("Disease Ontology",
                                   "GO Biological Process" ),
                                   exclude.type="")
man.Great.Object2 <- CombineGeneSets(Object = man.Great.Object1)
man.Great.Object3 <- ClusterGeneSets(Object = man.Great.Object2,
                                     clusters = 5,
                                     method = "kmeans")



}
