% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineGeneSets.R
\name{CombineGeneSets}
\alias{CombineGeneSets}
\title{CombineGeneSets}
\usage{
CombineGeneSets(Object, combineMethod = "Standard",
  combineMethod.supplied, display = "Condensed")
}
\arguments{
\item{Object}{A PathwayObject.}

\item{combineMethod}{lets the functions know if the standard RR, a Jaccard index or Cohens Kappa.}

\item{combineMethod.supplied}{a function which parameter A and parameter B.}

\item{display}{Either Condensed or Expanded. Expanded seperates the groups
and adds a 1 to every group is this molecular signature is seen here}
}
\value{
a pathwayobject
}
\description{
Calculate distances between the different experiments.
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
man.Great.Object1 <- ManageGeneSets(Object = Great.bckgnrd.Object1,
                                   keep.type =c("Disease Ontology",
                                   "GO Biological Process" ),
                                   exclude.type="")
man.Great.Object2 <- CombineGeneSets(Object = man.Great.Object1)

}
