% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetCluster.R
\docType{package}
\name{GeneSetCluster}
\alias{GeneSetCluster}
\title{GeneSetCluster package}
\description{
This is a package meant to be used to cluster together Gene-Sets from pathway tools to aid in biological interpretation.
}
\author{
Ewoud Ewing \email{ewoud.ewing@ki.se}
}
