% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesPerGeneSet.r
\name{GenesPerGeneSet}
\alias{GenesPerGeneSet}
\title{GenesPerGeneSet}
\usage{
GenesPerGeneSet(Object)
}
\arguments{
\item{Object}{a PathwayObject}
}
\value{
dataframe of pathwayobjects
}
\description{
Extracts a data.frame from the object with every gene in per cluster
}
