% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOptimalCluster.R
\name{GetOptimalCluster}
\alias{GetOptimalCluster}
\title{GetOptimalCluster}
\usage{
GetOptimalCluster(
  Object,
  cluster_method = "kmeans",
  max_cluster = 15,
  method = "gap"
)
}
\arguments{
\item{Object}{A PathwayObject}

\item{cluster_method}{kmeans of hclust}

\item{max_cluster}{A numeric with the number of maximal possible clusters}

\item{method}{The statistical method for determining the optimal number, currently only supports gap statistics}
}
\value{
k number of clusters
}
\description{
GetOptimalCluster
}
\examples{
GetOptimalGeneSetsClusters(Object, max_cluster = 6)
}
