% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSTRINGdbPerGeneSets.R
\name{GetSTRINGdbPerGeneSets}
\alias{GetSTRINGdbPerGeneSets}
\title{GetSTRINGdbPerGeneSets}
\usage{
GetSTRINGdbPerGeneSets(Object, unique.per.cluster = T,
  plot.input = "All")
}
\arguments{
\item{Object}{A PathwayObject.}

\item{unique.per.cluster}{A vector with strings of genes within each Gene-Set. Each string is seperated using the seperator supplied.}

\item{plot.input}{A vector with group names of the different gene set experiments}
}
\value{
a list with the string output and the prep for the plotting function PlotSTRINGdbPerGeneSets
}
\description{
GetSTRINGdbPerGeneSets
}
\examples{

IPA.files <- c(system.file("extdata", "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.KO.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files, #where are  the files
                           groupnames= c("KO", "WT"),#Names of the groups
                           P.cutoff = 1.3,
                           Mol.cutoff = 5,
                           Source = "IPA",
                           type = "Canonical_Pathways",
                           structure = "SYMBOL",
                           seperator = ",")
IPA.object2 <- CombineGeneSets(Object = IPA.object1)
IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                              clusters = 7,
                              method = "kmeans")
StringObject_unique <- GetSTRINGdbPerGeneSets(Object = IPA.object3,
                                             unique.per.cluster = TRUE ,
                                             plot.input = "All")
}
