% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighlightGeneSets.R
\name{HighlightGeneSets}
\alias{HighlightGeneSets}
\title{HighlightGeneSets}
\usage{
HighlightGeneSets(Object, highligt.genes, name = "Ros")
}
\arguments{
\item{Object}{A PathwayObject.}

\item{highligt.genes}{A vector with genes from the subset the user is interested in. e.g. a list of ROS genes.}

\item{name}{The name of the subset which will be added to the score calculated.}
}
\value{
a pathwayobject
}
\description{
Adds a highlight score if the Gene-Set overlaps with a gene subset which is supplied by the user.
}
\examples{
IPA.files <- c(system.file("extdata",
                           "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls",
                            package = "GeneSetCluster"),
             system.file("extdata",
                            "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.KO.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.WT.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files,
                         groupnames= c("KO", "WT"),
                         P.cutoff = 1.3,
                         Mol.cutoff = 5,
                         Source = "IPA",
                         type = "Canonical_Pathways",
                         structure = "SYMBOL",
                         seperator = ",")
IPA.object2 <- CombineGeneSets(Object = IPA.object1)

IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                              clusters = 12,
                              method = "kmeans")
system.file("data", "Redox.genes.rda", package = "testdat")
IPA.object3.highlight <- HighlightGeneSets(Object = IPA.object3,
                                          highligt.genes = Redox.genes,
                                          name = "Ros")

}
