% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGeneSets.R
\name{LoadGeneSets}
\alias{LoadGeneSets}
\title{LoadGeneSets}
\usage{
LoadGeneSets(file_location, groupnames = c("MClust_1", "MClust_26",
  "MClust_3457"), P.cutoff = 1.3, Mol.cutoff = 5, Source = "IPA",
  Great.Background = F, type = "Canonical_Pathways", topranks = "",
  structure = "SYMBOL", Organism = "org.Hs.eg.db", seperator = ",")
}
\arguments{
\item{file_location}{A location string in a vector.}

\item{groupnames}{A vector with group names of the different gene set experiments}

\item{P.cutoff}{numeric Pvalue cutoff for filtering.}

\item{Mol.cutoff}{numeric value for minimum number of molecules.}

\item{Source}{Tool used to generate gene sets.}

\item{Great.Background}{If the Great tool was used, did the user supply a background.}

\item{type}{For IPA data if Canonical pathways or Functional Anotations were supplied.}

\item{topranks}{numeric with the number of ranks per group to be loaded, usefull when there is a lot of data.}

\item{structure}{The structure of the genes. is it SYMBOLS, ENSEMBL, NCBI etc. Used for converting when there is mutiple structure in the object.}

\item{Organism}{the package name for the human or mouse data, used for converting the gene structure. name of the package, currently org.Hs.eg.db and org.Mm.eg.db supported.}

\item{seperator}{A character used within in the string to seperate genes}
}
\value{
a pathwayobject
}
\description{
Automatic loader for gene sets from the IPA and Great tools.
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")


}
