% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageGeneSets.R
\name{ManageGeneSets}
\alias{ManageGeneSets}
\title{ManageGeneSets}
\usage{
ManageGeneSets(Object, exclude.type = "", keep.type = "")
}
\arguments{
\item{Object}{A PathwayObject.}

\item{exclude.type}{A vector with strings of groups meant to be removed from the PathwayObject while keeping all others.}

\item{keep.type}{A vector with strings of groups meant to be kept from the PathwayObject removing all others.}
}
\value{
a pathwayobject
}
\description{
Filter unwanted groups from the PathwayObject
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
man.Great.Object1 <- ManageGeneSets(Object = Great.bckgnrd.Object1,
                                   keep.type =c("Disease Ontology",
                                   "GO Biological Process" ),
                                   exclude.type="")

}
