% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanClusterDifferences.R
\name{MeanClusterDifferences}
\alias{MeanClusterDifferences}
\title{BreakUpCluster}
\usage{
MeanClusterDifferences(Object)
}
\arguments{
\item{Object}{A PathwayObject}
}
\value{
PathwayObject
}
\description{
BreakUpCluster
}
