% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeObjects.R
\name{MergeObjects}
\alias{MergeObjects}
\title{MergeObjects}
\usage{
MergeObjects(Object.1, Object.2)
}
\arguments{
\item{Object.1}{A PathwayObject.}

\item{Object.2}{A PathwayObject.}
}
\value{
a pathwayobject
}
\description{
Merge 2 different PathwayObjects
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]
Great.files.NoBCKGRND <- Great.files[!grepl("BCKGRND", Great.files)]

Great.Object1 <- LoadGeneSets(file_location = Great.files.NoBCKGRND,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = FALSE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
GREAT.merged.object <- MergeObjects(Object.1 =Great.Object1,
                                   Object.2 = Great.bckgnrd.Object1)

}
