% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectCreator.R
\name{ObjectCreator}
\alias{ObjectCreator}
\title{ObjectCreator}
\usage{
ObjectCreator(Pathways, Molecules, Groups, Source, type, structure,
  Organism = NA, seperator)
}
\arguments{
\item{Pathways}{A vector with Gene-Set labels.}

\item{Molecules}{A vector with strings of genes within each Gene-Set. Each string is seperated using the seperator supplied.}

\item{Groups}{A vector with group names of the different gene set experiments}

\item{Source}{Tool used to generate gene sets..}

\item{type}{For IPA data if Canonical pathways or Functional Anotations were supplied.}

\item{structure}{The structure of the genes. is it SYMBOLS, ENSEMBL, NCBI etc. Used for converting when there is mutiple structure in the object.}

\item{Organism}{the package name for the human or mouse data, used for converting the gene structure. name of the package, currently org.Hs.eg.db and org.Mm.eg.db supported.}

\item{seperator}{A character used within in the string to seperate genes}
}
\value{
a pathwayobject
}
\description{
Manually create a PathwayObject
}
\examples{
Test.object <- matrix(data = NA, nrow = 50, ncol = 3)
colnames(Test.object) <- c("Pathways", "Genes", "Group")
Test.object[,"Pathways"] <- paste("Pathway", 1:nrow(Test.object),
                                 sep = "_")
Test.object[1:25,"Group"] <- "Group1"
Test.object[26:50,"Group"] <- "Group2"


#Create a random amount of genes per pathway
random.gene <- function()
{
 genenames <- paste("Gene", 1:200, sep = "_")#this gives 200 gene names
 genes <- round(runif(n = runif(n = 1,min =  7,max = 20),
                      min = 1, max = 200), digits = 0)
 #This gives between 7 and 20 random whole numbers
 genes <- unique(genes)#remove duplicate numbers
 genes <- genenames[genes]#get random genesnames
 return(genes)
}

for(i in 1:nrow(Test.object))
{
 Test.object[i, "Genes"] <- paste(random.gene(), collapse=",")
}
Test.object1 <- ObjectCreator(Pathways = Test.object[,1],
Molecules = Test.object[,2],
Groups = Test.object[,3],
Source = "Random",#we randomly generated this data
type = "Test",
structure = "SYMBOL",
sep = ",")#neccesay to seperate the different genes for combinations.




}
