% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalGeneSets.R
\name{OptimalGeneSets}
\alias{OptimalGeneSets}
\title{CombineGeneSets}
\usage{
OptimalGeneSets(
  Object,
  method,
  max_cluster,
  cluster_method,
  main,
  pathwayData = FALSE
)
}
\arguments{
\item{method}{Which method to determing optimal number of clusters. gap, elbow or silhouette.}

\item{max_cluster}{Max number of clusters to test}

\item{cluster_method}{kmeans or hcut. Which clustering method is used}

\item{main}{A string to be used as title in the plot}

\item{pathwayData}{Boolean to set Pathway or Cluster data}

\item{object}{A PathwayObject.}
}
\value{
a plot
}
\description{
Calculate distances between the different experiments.
}
