% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORAGeneSets.R
\name{ORAGeneSets}
\alias{ORAGeneSets}
\title{ORAGeneSets}
\usage{
ORAGeneSets(Object, ORA.returned = 10, unique.per.cluster = T)
}
\arguments{
\item{Object}{A PathwayObject.}

\item{ORA.returned}{A numeric for how Go terms should be returned per analysis. Alternative use "All" to get all of them.}

\item{unique.per.cluster}{A vector with strings of genes within each Gene-Set. Each string is seperated using the seperator supplied.}
}
\value{
a data frame with the over represented gene sets for all the clusters.
}
\description{
ORAGeneSets
}
\examples{

IPA.files <- c(system.file("extdata", "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.KO.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files, #where are  the files
                           groupnames= c("KO", "WT"),#Names of the groups
                           P.cutoff = 1.3,
                           Mol.cutoff = 5,
                           Source = "IPA",
                           type = "Canonical_Pathways",
                           structure = "SYMBOL",
                           seperator = ",")
IPA.object2 <- CombineGeneSets(Object = IPA.object1)
IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                              clusters = 7,
                              method = "kmeans")
ORAGeneSets(Object = IPA.object3,unique.per.cluster = TRUE ,
                                             ORA.returned = 1)
}
