% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayObject.R
\docType{class}
\name{PathwayObject-class}
\alias{PathwayObject-class}
\alias{PathwayObject}
\title{PathwayObject}
\value{
a PathwayObject
}
\description{
PathwayObject
}
\section{Slots}{

\describe{
\item{\code{Data}}{list.}

\item{\code{PData}}{data.frame.}

\item{\code{metadata}}{data.frame.}

\item{\code{plot}}{list.}

\item{\code{Data.RR}}{data.frame.}
}}

