% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClusterTissueExpression.R
\name{PlotClusterTissueExpression}
\alias{PlotClusterTissueExpression}
\title{PlotClusterTissueExpression}
\usage{
PlotClusterTissueExpression(Object, lclusters, all = FALSE)
}
\arguments{
\item{Object}{a PathwayObject}

\item{lclusters}{numeric vector giving which cluster(s) plot}

\item{all}{boolean to add all tissues}
}
\value{
dataframe of pathwayobjects with tissue expression per cluster
}
\description{
Plots the tissue expression of one or more clusters. Tissue expression is ordered based on the first cluster entered.
}
