% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneDedrogram.R
\docType{methods}
\name{PlotGeneDendrogram,PathwayObject-method}
\alias{PlotGeneDendrogram,PathwayObject-method}
\title{PlotGeneDendrogram}
\usage{
\S4method{PlotGeneDendrogram}{PathwayObject}(Object, load.order = F)
}
\arguments{
\item{Object}{a pathway object}

\item{load.order}{Boolean show to use the previous order of clusters, though this could result in non-hierarchical ordering of the dendrogram}
}
\value{
plot
}
\description{
PlotGeneDendrogram
}
\examples{

IPA.files <- c(system.file("extdata", "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls",
              package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls",
              package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.KO.uGvsMac.Functional_annotations.xls",
              package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.WT.uGvsMac.Functional_annotations.xls",
              package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files, #where are  the files
                           groupnames= c("KO", "WT"),
                           P.cutoff = 1.3,
                           Mol.cutoff = 5,
                           Source = "IPA",
                           type = "Canonical_Pathways",
                           structure = "SYMBOL",
                           seperator = ",")
 IPA.object2 <- CombineGeneSets(Object = IPA.object1)
 IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                            clusters = 7,
                            method = "kmeans")

 PlotGeneDendrogram(Object = IPA.object3, load.order=F)
}
