% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneNetworks.R
\docType{methods}
\name{PlotGeneNetworks,PathwayObject-method}
\alias{PlotGeneNetworks,PathwayObject-method}
\title{PlotGeneNetworks}
\usage{
\S4method{PlotGeneNetworks}{PathwayObject}(Object, labels = F,
  RRmin = 10)
}
\arguments{
\item{Object}{a pathway object}

\item{labels}{Boolean show labels or not, or a vector with labels to be shown}

\item{RRmin}{filter for different RR cutoff for what is an edge.}
}
\value{
plot
}
\description{
PlotGeneNetworks
}
