% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneNetworks.R
\name{PlotGeneNetworks}
\alias{PlotGeneNetworks}
\title{PlotGeneNetworks}
\usage{
PlotGeneNetworks(Object, labels = F, RRmin = 10)
}
\arguments{
\item{Object}{a pathway object}

\item{labels}{Boolean show labels or not, or a vector with labels to be shown}

\item{RRmin}{filter for different RR cutoff for what is an edge}
}
\value{
plot
}
\description{
Plots a heatmap with the distances per cluster
}
