% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneSets.R
\name{PlotGeneSets}
\alias{PlotGeneSets}
\title{PlotGeneSets}
\usage{
PlotGeneSets(Object, fontsize = 5, legend = T, annotation.mol = F,
  main = "", RR.max = "")
}
\arguments{
\item{Object}{a pathway object}

\item{fontsize}{a numeric with the fontsize for plotting}

\item{legend}{add a legend to the plot}

\item{annotation.mol}{should the genes from the genes set be added to the plot.}

\item{main}{is the plot title}

\item{RR.max}{is the maximum distance size to be added, it cutsoff the max to this. Usefull if the distance score gets very high.}
}
\value{
plot
}
\description{
Plots a heatmap with the distances per cluster
}
