% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLD.R
\name{PlotLD,PathwayObject-method}
\alias{PlotLD,PathwayObject-method}
\title{PlotLD}
\usage{
\S4method{PlotLD}{PathwayObject}(
  Object,
  fontsize = 8,
  annotation.fontsize = 3,
  main = "",
  label.triangle = ""
)
}
\arguments{
\item{Object}{a pathway object}

\item{fontsize}{a numeric with the fontsize for plotting rownames/colnames}

\item{annotation.fontsize}{a numeric with the fontsize for plotting the annotation labels}

\item{main}{is the plot title}

\item{label.triangle}{label name for plotting the triangles}
}
\value{
plot
}
\description{
PlotLD
}
\examples{
Great.files <- c(system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed.tsv",
                             package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.KO.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed.tsv", package = "GeneSetCluster"),
system.file("extdata", "MM10.GREAT.WT.uGvsMac.bed_BCKGRND.tsv", package = "GeneSetCluster"))
Great.files.bckgrnd <- Great.files[grepl("BCKGRND", Great.files)]


Great.bckgnrd.Object1 <- LoadGeneSets(file_location = Great.files.bckgrnd,
                                      groupnames= c("KO", "WT"),
                                      P.cutoff = 0.05,
                                      Mol.cutoff = 5,
                                      Source = "Great",
                                      Great.Background = TRUE,
                                      type = "Canonical_Pathways",
                                    topranks = 20,
                                   structure = "SYMBOL",
                                   Organism = "org.Mm.eg.db",
                                   seperator = ",")
man.Great.Object1 <- ManageGeneSets(Object = Great.bckgnrd.Object1,
                                   keep.type =c("Disease Ontology",
                                   "GO Biological Process" ),
                                   exclude.type="")
man.Great.Object2 <- CombineGeneSets(Object = man.Great.Object1)
man.Great.Object3 <- ClusterGeneSets(Object = man.Great.Object2,
                                     clusters = 5,
                                     method = "kmeans")
 PlotLD(Object = man.Great.Object3, fontsize =8,
              annotation.fontsize = 3,
              main="man.Great.Object3",
              label.triangle = "Cluster")
}
