% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLD.R
\name{PlotLD}
\alias{PlotLD}
\title{PlotLD}
\usage{
PlotLD(
  Object,
  fontsize = 8,
  annotation.fontsize = 3,
  main = "",
  label.triangle = ""
)
}
\arguments{
\item{Object}{a pathway object}

\item{fontsize}{a numeric with the fontsize for plotting rownames/colnames}

\item{annotation.fontsize}{a numeric with the fontsize for plotting the annotation labels}

\item{main}{is the plot title}

\item{label.triangle}{label name for plotting the triangles}
}
\value{
plot
}
\description{
Plots a LD with the distances per cluster
}
