% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTRINGdbPerGeneSets.R
\name{plotSTRINGdbPerGeneSets}
\alias{plotSTRINGdbPerGeneSets}
\title{plotSTRINGdbPerGeneSets}
\usage{
plotSTRINGdbPerGeneSets(StringObject, plot.cluster = 1)
}
\arguments{
\item{StringObject}{The output from GetSTRINGdbPerGeneSets}

\item{plot.cluster}{The cluster for which the string image should be plotted.}
}
\description{
plotSTRINGdbPerGeneSets
}
\examples{

IPA.files <- c(system.file("extdata", "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.KO.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"),
system.file("extdata", "MM10.IPA.WT.uGvsMac.Functional_annotations.xls", package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files, #where are  the files
                           groupnames= c("KO", "WT"),#Names of the groups
                           P.cutoff = 1.3,
                           Mol.cutoff = 5,
                           Source = "IPA",
                           type = "Canonical_Pathways",
                           structure = "SYMBOL",
                           seperator = ",")
IPA.object2 <- CombineGeneSets(Object = IPA.object1)
IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                              clusters = 7,
                              method = "kmeans")
StringObject_unique <- GetSTRINGdbPerGeneSets(Object = IPA.object3,
                                             unique.per.cluster = TRUE ,
                                             plot.input = "All")
par(mfrow=c(3,3))
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 1)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 2)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 3)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 4)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 5)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 6)
plotSTRINGdbPerGeneSets(StringObject = StringObject_all, plot.cluster = 7)
}
