% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTissueExpression.R
\name{PlotTissueExpression}
\alias{PlotTissueExpression}
\title{PlotTissueExpression}
\usage{
PlotTissueExpression(Object, all = FALSE)
}
\arguments{
\item{Object}{a pathway object}

\item{all}{boolean to add all tissues}
}
\value{
plot
}
\description{
Plots a the tissue expression total median expression per tissue and the percentage of each cluster in each tissue
}
