% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{PlotTree}
\alias{PlotTree}
\title{PlotTree}
\usage{
PlotTree(Object, fontsize = 3, main = "", nodenames = TRUE, wordcloud = TRUE)
}
\arguments{
\item{Object}{a pathway object}

\item{fontsize}{a numeric with the fontsize for plotting nodenames}

\item{main}{is the plot title}

\item{nodenames}{boolean to add names to the nodes}

\item{wordcloud}{boolean to add wordcloud annotations of each cluster}
}
\value{
plot
}
\description{
Plots a tree plot with the distances per cluster
}
