% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTreePathway.R
\name{PlotTreePathway,PathwayObject-method}
\alias{PlotTreePathway,PathwayObject-method}
\title{PlotTreePathway}
\usage{
\S4method{PlotTreePathway}{PathwayObject}(
  Object,
  clusters = 3,
  nodenames = TRUE,
  doORA = TRUE,
  wordcloud = TRUE
)
}
\arguments{
\item{Object}{a pathway object}

\item{clusters}{A numeric with the number of clusters required}

\item{nodenames}{boolean to add names to the nodes}

\item{doORA}{boolean perform ORA analysis}

\item{wordcloud}{boolean perform boolean to add wordcloud annotations of each cluster}
}
\value{
plot
}
\description{
PlotTreePathway
}
