% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Redox.genes}
\alias{Redox.genes}
\title{Redox.genes}
\format{
A vector with 513 gene names variables:
\describe{
  \item{genes}{gene names}
  ...
}
}
\source{
\url{http://www.informatics.jax.org/vocab/gene_ontology/GO:0000302}
}
\usage{
Redox.genes
}
\description{
A dataset genes from the GO term reactive oxygen species signaling
}
\keyword{datasets}
