% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowExperimentdata.R
\name{ShowExperimentdata}
\alias{ShowExperimentdata}
\title{ShowExperimentdata}
\usage{
ShowExperimentdata(Object)
}
\arguments{
\item{Object}{a PathwayObject}
}
\value{
a table with experiment data.
}
\description{
Show Experiment data
}
