% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowGeneSets.R
\name{ShowGeneSets}
\alias{ShowGeneSets}
\title{ShowGeneSets}
\usage{
ShowGeneSets(Object)
}
\arguments{
\item{Object}{A PathwayObject}
}
\value{
dataframe of pathwayobjects
}
\description{
SHows the Gene sets in the object
}
