% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowMeta.R
\name{ShowMeta}
\alias{ShowMeta}
\title{ShowMeta}
\usage{
ShowMeta(Object)
}
\arguments{
\item{Object}{a PathwayObject}
}
\value{
dataframe of pathwayobjects
}
\description{
shows experiment meta information
}
