% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPathwayCluster.R
\name{ShowPathwayCluster,PathwayObject-method}
\alias{ShowPathwayCluster,PathwayObject-method}
\title{ShowPathwayCluster}
\usage{
\S4method{ShowPathwayCluster}{PathwayObject}(Object, cluster = "ALL", nPathway = 4)
}
\arguments{
\item{Object}{a pathway object}

\item{nPathway}{minimum number of pathways per group}
}
\value{
data.frame
}
\description{
ShowPathwayCluster
}
