% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPathwayCluster.R
\name{ShowPathwayCluster}
\alias{ShowPathwayCluster}
\title{ShowPathwayCluster}
\usage{
ShowPathwayCluster(Object, cluster = "ALL", nPathway = 4)
}
\arguments{
\item{Object}{a pathway object}

\item{cluster}{the number of the cluster want to show information or "ALL" for all the clusters information.}

\item{nPathway}{minimum number of pathways per group, used in PlotPathwayCluster}
}
\value{
data.frame
}
\description{
Show the GO terms and descriptions of the generated cluster with PlotPathwayCluster.
}
