% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueExpressionPerGeneSet.R
\name{TissueExpressionPerGeneSet}
\alias{TissueExpressionPerGeneSet}
\title{TissueExpressionPerGeneSet}
\usage{
TissueExpressionPerGeneSet(Object, threads = 1)
}
\arguments{
\item{Object}{a PathwayObject}

\item{threads}{number of cores. Default the maximum of the computer.}

\item{PathwayObject}{a PathwayObject}
}
\value{
dataframe of pathwayobjects with tissue expression per cluster
}
\description{
Extracts every gen of every cluster from the object. Using the median gene expression from the GTEx database build a data.frame with every tissue per cluster.
}
