% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmapGeneSets.R
\name{UmapGeneSets}
\alias{UmapGeneSets}
\title{UmapGeneSets}
\usage{
UmapGeneSets(Object, main = "", labels = "")
}
\arguments{
\item{Object}{A PathwayObject}

\item{main}{Header for the plot}

\item{labels}{Labels of GeneSets to be highlighted}
}
\description{
UmapGeneSets
}
