% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{VennDiagram}
\alias{VennDiagram}
\title{VennDiagram}
\usage{
VennDiagram(
  n_groups = 4,
  Group1,
  Group2,
  Group3,
  Group4,
  names_groups,
  main,
  legend = F,
  percentage = F
)
}
\arguments{
\item{n_groups}{The number of groups to compare.}

\item{Group1}{A vector.}

\item{Group2}{A vector.}

\item{Group3}{A vector.}

\item{Group4}{A vector.}

\item{names_groups}{A vector with a names for each group.}

\item{main}{Plot title.}

\item{legend}{Add a legend to the plot.}

\item{percentage}{Add percentages to each group. Default is False}
}
\value{
VennDiagram
}
\description{
Plots a Venndiagram of overlap. Based on the limma functions.
}
\examples{
VennDiagram(n_groups = 2,
            Group1 =c("A","B","C","D","E"),
            Group2 =c("A","B","F","G","H"),
            names_groups = c("KO", "WT"),
            main = "Overlap",
            legend = FALSE,
            percentage = FALSE )

}
