% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteGeneSets.R
\name{WriteGeneSets}
\alias{WriteGeneSets}
\title{WriteGeneSets}
\usage{
WriteGeneSets(Object, file_location = "~/Project9/test/",
  name = "IPA_20181123", write = "Both")
}
\arguments{
\item{Object}{is a PathwayObject.}

\item{file_location}{where to write the files to}

\item{name}{name to be added to the files, what experiments are these}

\item{write}{what to write, either "Data", "RR" or "Both"}
}
\value{
written tables in the folder designated.
}
\description{
Exports the pathways into a csv file
}
\examples{
IPA.files <- c(system.file("extdata",
                           "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls",
                            package = "GeneSetCluster"),
             system.file("extdata",
                            "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.KO.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"),
             system.file("extdata",
                             "MM10.IPA.WT.uGvsMac.Functional_annotations.xls",
                             package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files,
                         groupnames= c("KO", "WT"),
                         P.cutoff = 1.3,
                         Mol.cutoff = 5,
                         Source = "IPA",
                         type = "Canonical_Pathways",
                         structure = "SYMBOL",
                         seperator = ",")
IPA.object2 <- CombineGeneSets(Object = IPA.object1)

IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                              clusters = 12,
                              method = "kmeans")
#WriteGeneSets(Object= IPA.object3,
 #           file_location = "~/Project9/test/",
 #            name = "IPA_20181123", write = "Both")


}
