% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeneNetworks.R
\name{PlotGeneNetworks,PathwayObject-method}
\alias{PlotGeneNetworks,PathwayObject-method}
\title{PlotGeneNetworks}
\usage{
\S4method{PlotGeneNetworks}{PathwayObject}(Object, labels = F, RRmin = 10)
}
\arguments{
\item{Object}{a pathway object}

\item{labels}{Boolean show labels or not, or a vector with labels to be shown}

\item{RRmin}{filter for different RR cutoff for what is an edge.}
}
\value{
plot
}
\description{
PlotGeneNetworks
}
\examples{

IPA.files <- c(system.file("extdata", "MM10.IPA.KO.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.WT.uGvsMac.Canonical_pathways.xls", package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.KO.uGvsMac.Functional_annotations.xls",package = "GeneSetCluster"),
              system.file("extdata", "MM10.IPA.WT.uGvsMac.Functional_annotations.xls",
              package = "GeneSetCluster"))
canonical.files <- IPA.files[grep("Canonical", IPA.files)]

IPA.object1 <- LoadGeneSets(file_location = canonical.files, #where are  the files
                           groupnames= c("KO", "WT"),
                           P.cutoff = 1.3,
                           Mol.cutoff = 5,
                           Source = "IPA",
                           type = "Canonical_Pathways",
                           structure = "SYMBOL",
                           seperator = ",")
 IPA.object2 <- CombineGeneSets(Object = IPA.object1)
 IPA.object3 <- ClusterGeneSets(Object = IPA.object2,
                            clusters = 7,
                            method = "kmeans")

 PlotGeneNetworks(Object = IPA.object3, labels=F, RRmin = 0)
}
