% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathwayCluster.R
\name{PlotPathwayCluster,PathwayObject-method}
\alias{PlotPathwayCluster,PathwayObject-method}
\title{PlotPathwayCluster}
\usage{
\S4method{PlotPathwayCluster}{PathwayObject}(Object, nPathway = 6, doORA = TRUE, wordcloud = TRUE)
}
\arguments{
\item{Object}{a pathway object}

\item{nPathway}{minimum number of pathways per group}

\item{doORA}{boolean perform ORA analysis}
}
\value{
plot
}
\description{
PlotPathwayCluster
}
