% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathwayCluster.R
\name{PlotPathwayCluster}
\alias{PlotPathwayCluster}
\title{PlotPathwayCluster}
\usage{
PlotPathwayCluster(Object, nPathway = 4)
}
\arguments{
\item{Object}{a pathway object}

\item{nPathway}{minimum number of pathways per group}
}
\value{
plot
}
\description{
Plots a correlation matrix showing the correlation between the pathways and a functional annotation of the group of pathways based on GO terms descriptions.
}
